fix(stdlib): fix issues when using LVGL TLSF memory pool to manage more than 4 GiB of memory (#5720)

+++ tests/src/test_cases/cache/test_cache.c
@@ -127,7 +127,7 @@ void test_cache_1(void)
      * Now drop the cache {key1 = 32, key2 = 33}. However, this entry is acquired once without release, so `drop`
      * will not release the memory allocated by this entry.
      */
-    uint32_t mem_curr_free = lv_test_get_free_mem();
+    size_t mem_curr_free = lv_test_get_free_mem();
     lv_cache_drop(cache, &search_key32, NULL);
     /*
      * Though it doesn't release the data, the entry and other structure has been freed.
